requirements.txt
streamlit
pandas
fpdf
python-dotenv
crewai